Move small scenery functions to cpp file
